Fix "no matching function" error on yocto kirkstone (#331)
### Changelog
None
### Docs
None
### Description

#330 Fixes for Yocto issue

Fix for compilation error on yocto Kirkstone.

> error: no matching function for call to 'max(long unsigned int, size_t)

+++ ros2_foxglove_bridge/src/ros2_foxglove_bridge.cpp
@@ -480,7 +480,7 @@ void FoxgloveBridge::subscribe(foxglove::ChannelId channelId, ConnectionHandle c
     // broadcasters). See also
     // https://github.com/foxglove/ros-foxglove-bridge/issues/238 and
     // https://github.com/foxglove/ros-foxglove-bridge/issues/208
-    const size_t publisherHistoryDepth = std::max(1ul, qos.depth());
+    const size_t publisherHistoryDepth = std::max(static_cast<size_t>(1), qos.depth());
     depth = depth + publisherHistoryDepth;
   }
